requirements.txt
pypdf
scikit-learn
pandas 
numpy 
pyresparser
tensorflow
keras
nltk                        
streamlit 
rapidfuzz
langchain
langchain_community
python-dotenv
langchain-google-genai
pandas
langchain-astradb
streamlit
tiktoken
langgraph
langchainhub
pypdf
easyocr
langchain-astradb
tiktoken
langchain-core
langchain-astradb
spacy
groq
pymongo
seaborn